benchmarks/python/requirements.txt
# ANN Benchmark Dependencies
faiss-cpu>=1.7.4
hnswlib>=0.8.0
annoy>=1.17.3

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
psutil>=5.9.0
requests>=2.31.0
tabulate>=0.9.0
